refactor(api): split ingest record readers into record_reader.go
Extract the ~210-line format-detection / record-reader machinery (the
recordReader interface, the object/array/line readers, the sentinels, and
the content-sniffing newRecordReader factory) out of ingest.go into a new
record_reader.go. ingest.go drops from 665 to 428 lines and is now focused
on the handler pipeline (Handle / handleSingle / handleBatch /
processRecord). No behavior change.

Also DRY the whole-request abort response (status + optional Retry-After)
into a shared writeAbort helper used by both the single-object and batch
paths.

Revert the README ingest array example (those docs land on a separate
branch) and drop README.md from the CHANGELOG file list; add
record_reader.go there and point AGENTS.md design decision #3 at it.
