+++ AGENTS.md
@@ -30,7 +30,7 @@ Eleven internal packages under `internal/` (plus `internal/testutil/` for shared
 
 1. **Interface-first**: Core behaviors are defined as Go interfaces (`Cache`, `Deduplicator`, `Publisher`, `Subscriber`). Standalone and (future) clustered modes use different implementations.
 2. **Bring Your Own Schema (BYOS)**: Users create tables in ClickHouse directly. WaveHouse discovers schemas by querying `system.columns` and validates ingest payloads against real column definitions. No auto-migration, no fixed table schema.
-3. **Schema-driven ingest**: `POST /v1/ingest?table={table}` accepts a flat JSON object, a JSON array of objects, or an NDJSON batch. The table name comes from the `table` query parameter. The body is validated against the discovered schema (unknown fields rejected, types checked, nullable constraints enforced). No envelope — just data. The **format is auto-detected from the body — the `Content-Type` is only a hint** (the first non-whitespace byte decides; an explicit `application/x-ndjson` selects line-framing unless the body starts with `[`). A batch (array/NDJSON) fans into one independently-validated/published message per record and returns a per-record `{total,succeeded,failed,duplicates,results:[{index, ok|duplicate|error}]}` summary, so one malformed/rejected record never fails the rest of the batch; a lone object keeps its `{"ok":true}` / `{"duplicate":true}` response. Whole-request conditions — table-level `insert` denial, a JSON-array syntax error / oversized NDJSON line (`400`), the 16 MiB body cap (`413`, parity with `/v1/admin/query`), backpressure (`503`), publish/dedup-backend errors (`500`) — still abort non-200. Formats sit behind a pull-based `recordReader` seam (object / array / line readers + a body sniffer) in `internal/api/ingest.go` so CSV and response-streaming slot in later; the single and batch paths share the `processRecord` helper. Bounded-concurrency uploading of large batches is deferred to #196.
+3. **Schema-driven ingest**: `POST /v1/ingest?table={table}` accepts a flat JSON object, a JSON array of objects, or an NDJSON batch. The table name comes from the `table` query parameter. The body is validated against the discovered schema (unknown fields rejected, types checked, nullable constraints enforced). No envelope — just data. The **format is auto-detected from the body — the `Content-Type` is only a hint** (the first non-whitespace byte decides; an explicit `application/x-ndjson` selects line-framing unless the body starts with `[`). A batch (array/NDJSON) fans into one independently-validated/published message per record and returns a per-record `{total,succeeded,failed,duplicates,results:[{index, ok|duplicate|error}]}` summary, so one malformed/rejected record never fails the rest of the batch; a lone object keeps its `{"ok":true}` / `{"duplicate":true}` response. Whole-request conditions — table-level `insert` denial, a JSON-array syntax error / oversized NDJSON line (`400`), the 16 MiB body cap (`413`, parity with `/v1/admin/query`), backpressure (`503`), publish/dedup-backend errors (`500`) — still abort non-200. Formats sit behind a pull-based `recordReader` seam (object / array / line readers + a body sniffer) in `internal/api/record_reader.go` so CSV and response-streaming slot in later; the single and batch paths in `internal/api/ingest.go` share the `processRecord` helper. Bounded-concurrency uploading of large batches is deferred to #196.
 4. **Async ingestion**: Ingest returns 200 immediately after optional dedup + MQ publish. ClickHouse writes happen asynchronously via the ingest worker pipeline (`StartIngestWorker`). If NATS stream is full, returns 503 + Retry-After.
 5. **Per-table batching**: The ingest worker pipeline groups events by table name and performs dynamic INSERTs using the schema's column order. Each table's batch is independent.
 6. **Dead Letter Queue**: Failed batch inserts are published to a separate NATS stream (`WAVEHOUSE_DLQ`) with subjects `dlq.<table>`. This prevents silent data loss. Controlled by `dlq.enabled`.
