+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- **Forgiving multi-format ingest on `POST /v1/ingest` (single object, JSON array, or NDJSON), plus SDK `insert([...])` / `insertNDJSON()`** (`internal/api/ingest.go`, `internal/api/ingest_test.go`, `clients/ts/src/{http,types,table,index}.ts`, `clients/ts/src/{http,table}.test.ts`, `tests/e2e/sdk/ndjson.test.ts` (new), `tests/e2e/sdk/tables.ts`, `docs/src/content/docs/{api,sdk}.md`, `README.md`, `AGENTS.md`): closes #195 (part of #194). The ingest endpoint now accepts a single flat JSON object, a **JSON array** of objects (any length, even one), or an **NDJSON** batch — the format is auto-detected from the body, with `Content-Type` only a hint (the first non-whitespace byte decides; an explicit `application/x-ndjson` selects line-framing unless the body starts with `[`). Each record is validated, authorized, deduplicated, and published independently, so one malformed or rejected record never blocks the rest of the batch; a batch (array/NDJSON) returns a per-record summary `{total, succeeded, failed, duplicates, results:[{index, ok|duplicate|error}]}` with `200` whenever the body was readable (branch on `failed`/`results`, not the status), while a lone object keeps its `{"ok":true}` / `{"duplicate":true}` response byte-for-byte. Whole-request conditions abort with a non-`200`: a table-level `insert` denial (`403`, checked once up front), a JSON-array syntax error or oversized NDJSON line (`400`), the new 16 MiB body cap (`413`, parity with `/v1/admin/query`), backpressure (`503` + `Retry-After`), and publish/dedup-backend errors (`500`). Formats sit behind a pull-based `recordReader` seam (object / array / line readers + a content sniffer) so CSV and response-streaming can drop in later without reworking the handler; per-record vs whole-request handling stays in the shared `processRecord` helper.
+- **Forgiving multi-format ingest on `POST /v1/ingest` (single object, JSON array, or NDJSON), plus SDK `insert([...])` / `insertNDJSON()`** (`internal/api/ingest.go`, `internal/api/record_reader.go` (new), `internal/api/ingest_test.go`, `clients/ts/src/{http,types,table,index}.ts`, `clients/ts/src/{http,table}.test.ts`, `tests/e2e/sdk/ndjson.test.ts` (new), `tests/e2e/sdk/tables.ts`, `docs/src/content/docs/{api,sdk}.md`, `AGENTS.md`): closes #195 (part of #194). The ingest endpoint now accepts a single flat JSON object, a **JSON array** of objects (any length, even one), or an **NDJSON** batch — the format is auto-detected from the body, with `Content-Type` only a hint (the first non-whitespace byte decides; an explicit `application/x-ndjson` selects line-framing unless the body starts with `[`). Each record is validated, authorized, deduplicated, and published independently, so one malformed or rejected record never blocks the rest of the batch; a batch (array/NDJSON) returns a per-record summary `{total, succeeded, failed, duplicates, results:[{index, ok|duplicate|error}]}` with `200` whenever the body was readable (branch on `failed`/`results`, not the status), while a lone object keeps its `{"ok":true}` / `{"duplicate":true}` response byte-for-byte. Whole-request conditions abort with a non-`200`: a table-level `insert` denial (`403`, checked once up front), a JSON-array syntax error or oversized NDJSON line (`400`), the new 16 MiB body cap (`413`, parity with `/v1/admin/query`), backpressure (`503` + `Retry-After`), and publish/dedup-backend errors (`500`). Formats sit behind a pull-based `recordReader` seam (object / array / line readers + a content sniffer) so CSV and response-streaming can drop in later without reworking the handler; per-record vs whole-request handling stays in the shared `processRecord` helper.
   - **SDK.** `from(t).insert(rows[])` now serializes to a **single** NDJSON request instead of firing one unbounded concurrent POST per row (which also silently returned only the first error). Per-record outcomes surface on `InsertResult` (`total`/`succeeded`/`failed`/`duplicates`/`results`), and `ok` is `true` only when every record succeeded; the call's top-level `error` is reserved for whole-request failures. A new `from(t).insertNDJSON(source)` sends NDJSON you already have (`string` / `Uint8Array` / `Blob`/`File` / `ReadableStream`). An empty array is a no-op (no request). New exported types `InsertRecordResult` and `NDJSONSource`.
   - **Deferred to #196:** bounded-concurrency chunking of very large arrays + backpressure-aware retry. `insert([...])` currently sends one request regardless of size.
 - **CORS support for the SSE streaming endpoint** (`internal/api/router.go`, `internal/api/stream.go`, `internal/api/router_test.go`, `docs/src/content/docs/api.md`): closes #215. `GET /v1/stream` already inherited the router-level `corsMiddleware` (`Access-Control-Allow-Origin` echo + OPTIONS preflight), but the SSE resumption header `Last-Event-ID` — which the handler reads to gap-fill from the client's last seen event — was not in `Access-Control-Allow-Headers`, so a cross-origin fetch-based SSE client resuming via that header failed its preflight. `Last-Event-ID` is now allow-listed, the SSE handler carries a comment pointing at the middleware, and new end-to-end tests pin CORS on `/v1/stream` (preflight advertises `Last-Event-ID`, the streaming GET echoes an allowed origin, a disallowed origin gets no header) so the behavior can't silently regress. Browser `EventSource` clients still pass the JWT via `?token=` since they can't set an `Authorization` header.
